Add files via upload
The individual pdf files contain the raw pycnometry data related to 4 different Agarose concentrations (A0.5, A0.75, A1 and A2) subjected to two different drying methods (vacuum drying and freeze drying). The designation 5/30 indicates that the measurements or evaluations were conducted at the 72-hour timepoint (i.e., 72 hours after the beginning of the drying on 5/27 ). Related images and plots are reflected in the Summary.pdf file in this repository.